Added a sources page. Added images to history.html
At the bottom of history.html (footer section), there is now a line
directing users to a source page if they wish to see it.

The source page contains a title of a page on the website followed by hyperlinks in a
list format to sources used to help create the page. The main content is
split into two sections, one side for generic sources, and the other for
image sources.

If you would like to add your sources, feel free to add to the
sources.html file. Don't forget to modify your footer (use history.html
as an example) to direct users to that page!