Fixes issue #19. Also properly centers titles.
The anchor tags on the sides of the body have been removed. I'm still
confused on how they were there. It would make sense if <a> was a block
element, but it's not (as far as I know). And it was only clickable on
some parts of the body sides.

I fixed the issue by targeting the title anchor tag and setting it to
display: block; Then I was able to resize it to fit the title box.

Also, the title box is now properly centered in the header. I had to
manually center it by finding the correct amount of pixels to push it
from the top using margin-top. For some reason, there was extra space on
the bottom, and I still don't know why.